0.1.0: Removing cruft, final archived release.
This was a long journey basing various build systems around tup.

At this point this tool is retired. This release removes a lot of
cruft related to 'server' support that never quite matured beyond
basic shell-esque scripting.

The incremental build features from tup, the explicit inlining and
the exposed and permanently in-tree build script access allowed for
a very precise and flexible setup that had proven adaptable in
projects over the years.

The primary drawback is that the developer experience simply pales
in comparison to Vite. Furthermore, the middle-ground between
Vue/Vuex/VueRouter and stage0/S.js seems to be Solid.js. Ryan
certainly built a very nice toolkit with some great influences.

Between all of this, it seems more prudent to spend time bringing
ideas from here to the Vite/Solid/Tailwind pipeline.